requirements.txt
alembic==1.0.11
bcrypt==3.1.7
blinker==1.4
cachetools==4.1.0
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
dominate==2.4.0
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.9
Flask-CSV==1.2.0
Flask-Executor==0.9.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-User==0.6.21
Flask-WTF==0.14.3
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
gspread==3.6.0
gevent==20.9.0
greenlet==0.4.17
grequests==0.6.0
httplib2==0.18.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.9.0
numpy==1.17.3
oauth2client==4.1.3
oauthlib==3.1.0
pandas==0.24.2
passlib==1.7.1
protobuf==3.12.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.0
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.2
six==1.15.0
SQLAlchemy==1.3.20
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.11
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.2.1
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.2.0
